refactor(build-api): rename taskStatus.ts to taskStatusUtils.ts (#1185)
File contains only utility functions (normalizeProgressPhase, etc.)
but did not use the *Utils.ts suffix per naming guideline.

Rollback: git mv packages/build-api/src/taskStatusUtils.ts packages/build-api/src/taskStatus.ts
