+++ packages/build-api/src/index.ts
@@ -1,6 +1,6 @@
 export * from './isBuildControlAPIV2Enabled.js';
 export * from './progressPayloadUtils.js';
-export * from './taskStatus.js';
+export * from './taskStatusUtils.js';
 export type {
   BuildStage,
   BuildContinuationPolicy,
